+++ src/components/About us/usercard.css
@@ -1,0 +1,128 @@
+.container2 {
+    width: 100%;
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.container2 .card {
+    position: relative;
+    border-radius: 10px;
+    width: 70vw;
+}
+
+.container2 .card .icon {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: #f00;
+    transition: 0.7s;
+    z-index: 1;
+}
+
+.container2 .card:nth-child(1) .icon {
+    background: #e07768;
+}
+
+.container2 .card:nth-child(2) .icon {
+    background: #6eadd4;
+}
+
+.container2 .card:nth-child(3) .icon {
+    background: #4aada9;
+}
+
+
+.container2 .card .icon .fa {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -150%);
+    font-size: 80px;
+    transition: 0.7s;
+    color: #fff;
+}
+
+i {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -150%);
+    font-size: 80px;
+    transition: 0.7s;
+    color: #fff;
+}
+
+.container2 .card .face {
+    width: 100%;
+    height: 300px;
+    transition: 0.5s;
+}
+
+.container2 .card .face.face1 {
+    position: relative;
+    background: #333;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1;
+    transform: translateY(200px);
+}
+
+.container2 .card:hover .face.face1{
+    background: #ff0057;
+    transform: translateY(0px);
+}
+
+.container2 .card .face.face1 .content {
+    opacity: 1;
+    transition: 0.5s;
+}
+
+.container2 .card:hover .face.face1 .content {
+    opacity: 1;
+}
+
+.container2 .card .face.face1 .content i{
+    max-width: 100px;
+}
+
+.container2 .card .face.face2 {
+    position: relative;
+    background: #fff;
+    display: flex;
+    justify-content: center;
+    padding: 10px;
+    box-sizing: border-box;
+    box-shadow: 0 20px 50px rgba(0,0,0,0.8);
+    transform: translateY(-100px);
+}
+
+.container2 .card:hover .face.face2{
+    transform: translateY(0);
+}
+
+.container2 .card .face.face2 .content p {
+    margin: 0;
+    padding: 0;
+    text-align: center;
+    color: #414141;
+}
+
+.container2 .card .face.face2 .content h3 {
+    margin: 0 0 10px 0;
+    padding: 0;
+    color: #fff;
+    font-size: 24px;
+    text-align: center;
+    color: #414141;
+} 
+
+a{
+    text-decoration: none;
+    color: white;
+}
